Simplified transport tests
