+++ src/test/java/io/vertx/ext/mongo/tests/EpollTransportTest.java
@@ -1,0 +1,13 @@
+package io.vertx.ext.mongo.tests;
+
+import static io.vertx.core.transport.Transport.EPOLL;
+
+import io.vertx.core.transport.Transport;
+
+public class EpollTransportTest extends NativeTransportTestBase {
+
+  @Override
+  protected Transport vertxTransport() {
+    return EPOLL;
+  }
+}
